+++ R/docinfo.R
@@ -1,6 +1,5 @@
 #' Set/get pdf document info dictionary
 #'
-#' `docinfo()` creates an object of pdf document info dictionary information from scratch.
 #' `get_docinfo()` gets pdf document info from a file.
 #' `set_docinfo()` sets pdf document info for a file.
 #'
@@ -47,7 +46,7 @@
 #'   grid.text("Page 2")
 #'   invisible(dev.off())
 #'
-#'   cat("\nInitial documentation info:\n")
+#'   cat("\nInitial documentation info:\n\n")
 #'   print(get_docinfo(f)[[1]])
 #'
 #'   d <- docinfo(author = "John Doe",
@@ -56,7 +55,7 @@
 #'                filename = f)
 #'   set_docinfo(d, f)
 #'
-#'   cat("\nDocumentation info after setting it:\n")
+#'   cat("\nDocumentation info after setting it:\n\n")
 #'   print(get_docinfo(f)[[1]])
 #'
 #'   unlink(f)